+++ packages/zowe-explorer-api/CHANGELOG.md
@@ -38,7 +38,66 @@ All notable changes to the "zowe-explorer-api" extension will be documented in t
 - **Next Breaking:** Changed `ProfilesCache.convertV1ProfToConfig` method to be a static method that requires `ProfileInfo` instance as a parameter.
 - Added the `onVaultUpdate` VSCode event to notify extenders when credentials are updated on the OS vault by other applications. [#2994](https://github.com/zowe/zowe-explorer-vscode/pull/2994)
 - Added the `onCredMgrsUpdate` VSCode event to notify extenders when the local PC's credential manager has been updated by other applications. [#2994](https://github.com/zowe/zowe-explorer-vscode/pull/2994)
-- **Breaking:** Updated most function signatures for exported programmatic interfaces. Changes make developing with the Zowe Explorer API more efficient for extenders by showing which properties they can expect when calling our APIs. [#2952](https://github.com/zowe/zowe-explorer-vscode/issues/2952)
+- **LTS Breaking:** Updated most function signatures for exported programmatic interfaces. Changes make developing with the Zowe Explorer API more efficient for extenders by showing which properties they can expect when calling our APIs. [#2952](https://github.com/zowe/zowe-explorer-vscode/issues/2952)
+
+  - Updated `IApiExplorerExtender.ts`, see changes below:
+    - Allowed `reloadProfiles` and `initForZowe` to be synchronous methods (non-breaking)
+  - Updated `MainframeInteraction.ts`, see changes below:
+    - Modified `getStatus` to add `string` type to the optional parameter `profileType`
+  - Updated `IZoweTree.ts`, see changes below:
+
+    - Modified `checkCurrentProfile(node: IZoweTreeNode);` to return `Validation.IValidationProfile | Promise<Validation.IValidationProfile>`
+    - Modified `getSearchHistory()` to return `string[]`
+    - Modified `getAllLoadedItems()` to return `IZoweTreeNode[] | Promise<IZoweTreeNode[]>`
+    - Modified `getFileHistory()` to return `string[]`
+    - Modified `createFilterString?(newFilter: string, node: IZoweTreeNode);` to return `string`
+    - Allowed the following methods to be implemented synchronously: (non-breaking)
+      - `addSession`
+      - `addSingleSession`
+      - `editSession`
+      - `createZoweSession`
+      - `createZoweSchema`
+      - `ssoLogin`
+      - `ssoLogout`
+      - `addFavorite`
+      - `removeFavorite`
+      - `removeFavProfile`
+      - `onDidChangeConfiguration`
+      - `flipState` _synchronous - no `Promise` returned_
+      - `rename`
+      - `open`
+      - `copy`
+      - `paste`
+      - `delete`
+      - `setItem` _synchronous - no `Promise` returned_
+      - `saveSearch`
+      - `saveFile`
+      - `refreshPS`
+      - `uploadDialog`
+      - `filterPrompt`
+      - `addSearchHistory` _synchronous - no `Promise` returned_
+      - `deleteSession` _synchronous - no `Promise` returned_
+      - `updateFavorites`
+      - `renameFavorite`
+      - `addFileHistory` _synchronous - no `Promise` returned_
+      - `removeFileHistory` _synchronous - no `Promise` returned_
+      - `renameNode`
+      - `openItemFromPath`
+      - `addDsTemplate`
+      - `pollData`
+      - `openWithEncoding`
+
+    **Note**: Developers should not expect a value to be returned from the methods above (breaking)
+
+  - Updated `IZoweTreeNode.ts`, see changes below:
+    - Modified `rename?(newNamePath: string);` to return `Promise<IZosFilesResponse>`
+    - Developers should not be expecting output from the following methods:
+      - `openUSS`
+      - `refreshUSS`
+      - `deleteUSSNode`
+      - `renameUSSNode`
+      - `reopen`
+      - `saveSearch`
 
 ### Bug fixes